fix: unblock markdown alerts CI and correct nested section counts
Split non-component parsers out of React files for fast-refresh budget,
fix blockquote marker typing, and stop counting sections nested inside
alerts or plain quotes as direct children.
